Refactorizacion de tests
