--- src/customers/customers.service.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { CustomersService } from './customers.service';
-
-describe('CustomersService', () => {
-  let service: CustomersService;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      providers: [CustomersService],
-    }).compile();
-
-    service = module.get<CustomersService>(CustomersService);
-  });
-
-  it('should be defined', () => {
-    expect(service).toBeDefined();
-  });
-});
