+++ src/customers/customers.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { CreateCustomerDto } from './dto/create-customer.dto';
 import { UpdateCustomerDto } from './dto/update-customer.dto';
-import { Customer } from './entities/customer.entity';
+import { Customer } from './domain/customer.entity';
 import { DB_CUSTOMERS } from './dbCustomers.const';
 
 @Injectable()
