+++ src/service-orders/entities/interfaces/coordinate.interface.ts
@@ -1,4 +1,4 @@
-export interface Coordinate {
+export type Coordinate = {
   latitude: number;
   longitude: number;
-}
+};
